requirement.txt
flask
mysql-connector-python
pandas
sklearn
numpy
pickle
logging
